Add Contact to the nav-bar and update react route
