+++ src/App.jsx
@@ -1,34 +1,35 @@
-import { useState } from 'react'
-import Navbar from './components/Navbar'
+import TodoWrapped from './components/TodosLogic';
+import Root from './routes/root';
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import ErrorPage from './routes/error-page';
+import Contact from './routes/contact';
 
-function App() {
-  const [count, setCount] = useState(0)
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Root />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        path: "contact/:contactId",
+        element: <Contact />,
+      },
+      {
+        path: "/",
+        element: <TodoWrapped />
+        
+      },
+      
+    ],
+  },
+]);
 
+const TodoApp = () => {
   return (
     <>
-    <Navbar />
-      <div>
-        <a href="https://vitejs.dev" target="_blank" rel="noreferrer">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank" rel="noreferrer">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
+      <RouterProvider router={router}/>        
+      
     </>
-  )
-}
-
-export default App
+  );
+};
+export default TodoApp;
