+++ src/components/InputTodo.jsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import { FaPlusCircle } from "react-icons/fa"
+import PropTypes from 'prop-types'
 
 const InputTodo = ({ addTodoItem }) => {
   const [title, setTitle] = useState('');
@@ -38,4 +39,9 @@ const InputTodo = ({ addTodoItem }) => {
     </>
   );
 };
+
+InputTodo.propTypes = {
+  addTodoItem: PropTypes.func.isRequired,
+}
+
 export default InputTodo;
