+++ src/components/TodoList.jsx
@@ -1,16 +1,32 @@
-import TodoItem from "./TodoItem";
+import TodoItem from './TodoItem';
+import PropTypes from 'prop-types';
+
 const TodosList = ({ todosProps, handleChange, delTodo, setUpdate }) => {
   return (
     <ul>
       {todosProps.map((todo) => (
-        <TodoItem 
-        key={todo.id} 
-        itemProp={todo} 
-        handleChange={handleChange}
-        delTodo={delTodo} 
-        setUpdate={setUpdate} />
+        <TodoItem
+          key={todo.id}
+          itemProp={todo}
+          handleChange={handleChange}
+          delTodo={delTodo}
+          setUpdate={setUpdate}
+        />
       ))}
     </ul>
   );
 };
+
+TodosList.propTypes = {
+  todosProps: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      // Specify other properties of the todo item here
+    })
+  ).isRequired,
+  handleChange: PropTypes.func.isRequired,
+  delTodo: PropTypes.func.isRequired,
+  setUpdate: PropTypes.func.isRequired
+};
+
 export default TodosList;
