+++ src/main.jsx
@@ -2,7 +2,7 @@ import './App.css'
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import TodoApp from './components/TodoApp';
+import TodoApp from './App';
 
 const domContainer = document.getElementById('root');
 const root = ReactDOM.createRoot(domContainer);
