+++ src/styles/Navbar.module.css
@@ -41,7 +41,6 @@
 
 a {
   text-decoration: none;
-
 }
 
 a:visited {
